projects/core/gcp-oece-tech/data/legacy/repos/deepweay/apps/telegram-bot/requirements-fixed.txt
# Telegram Bot Core
python-telegram-bot==20.7

# 異步 HTTP
aiohttp==3.9.1
aiofiles==23.2.1

# Redis
redis==5.0.1

# 數據庫（使用二進制版本，不需要編譯）
psycopg2-binary==2.9.9

# Supabase
supabase==2.3.4

# 環境管理
python-dotenv==1.0.0

# HTTP 請求
requests==2.31.0

# 日誌
loguru==0.7.2
